Fix mistyped ngbDropdownMenu in some ngbDropdown ul elements.
Some dropdown menu did not show up, because ngbDropdown directive was used in the ul element instead of ngbDropdownMenu.

This issue appeared in the notification component page, and in the traffic section in the dashboard.

It works in the live demo page, but did not work in my localhost before this correction.